vir & shaderthing - plethora of small changes to address most of the compiler warnings which I have conveniently ingored so far | shaderthing - wrapped overall app execution in try catch to try and log something, anything, when crashing in release (miserable attempt but will do for now); improved safety of FileDialog by introducing additional checks on the status of task_, selection_; reduced release optimization level from O3 to O2

+++ shaderthing/src/objectio.cpp
@@ -428,7 +428,7 @@ std::vector<std::string> ObjectIO::read(const char* key) const
     ASSERT_READ_MODE_OR_RETURN(std::vector<std::string>(0))
     auto a = ((nativeReader*)nativeObject_)->operator[](key).GetArray();
     std::vector<std::string> v(a.Size());
-    for (int i=0; i<a.Size(); i++)
+    for (int i=0; i<(int)a.Size(); i++)
     {
         v[i] = a[i].GetString();
     }
@@ -441,7 +441,7 @@ std::vector<const char*> ObjectIO::read(const char* key) const
     ASSERT_READ_MODE_OR_RETURN(std::vector<const char*>(0))
     auto a = ((nativeReader*)nativeObject_)->operator[](key).GetArray();
     std::vector<const char*> v(a.Size());
-    for (int i=0; i<a.Size(); i++)
+    for (int i=0; i<(int)a.Size(); i++)
     {
         // Force copy
         unsigned int srcSize = a[i].GetStringLength();
@@ -464,7 +464,7 @@ std::vector<const char*> ObjectIO::readOrDefault
     CHECK_KEY
     auto a = ((nativeReader*)nativeObject_)->operator[](key).GetArray();
     std::vector<const char*> v(a.Size());
-    for (int i=0; i<a.Size(); i++)
+    for (int i=0; i<(int)a.Size(); i++)
     {
         // Force copy
         unsigned int srcSize = a[i].GetStringLength();
